refactor: move web search utility functions to a shared module for reuse across extension and preferences
