+++ prefs.js
@@ -10,7 +10,7 @@ import Gdk from 'gi://Gdk';
 import Adw from 'gi://Adw';
 import GLib from 'gi://GLib';
 import Gio from 'gi://Gio';
-import Soup from 'gi://Soup?version=3.0';
+import { getIconPath, fetchFaviconForProvider } from './webSearchUtils.js';
 
 function buildPrefsWidget() {
   let provider = new Gtk.CssProvider();
@@ -417,64 +417,6 @@ function getProvidersFromSettings(settings) {
 
 function saveProviders(settings, providers) {
   settings.set_string('web-search-providers', JSON.stringify(providers));
-}
-
-function getIconsDir() {
-  return GLib.build_filenamev([GLib.get_user_cache_dir(), 'switcher', 'icons']);
-}
-
-function getIconPath(keyword) {
-  return GLib.build_filenamev([getIconsDir(), `${keyword}.png`]);
-}
-
-function extractDomain(url) {
-  const match = url.match(/^https?:\/\/([^\/]+)/);
-  if (!match) return null;
-  return match[1].replace(/^www\./, '');
-}
-
-function fetchFavicon(url, outputPath, callback) {
-  try {
-    const session = new Soup.Session();
-    const message = Soup.Message.new('GET', url);
-    session.send_and_read_async(message, GLib.PRIORITY_DEFAULT, null, (source, result) => {
-      try {
-        const bytes = session.send_and_read_finish(result);
-        if (message.get_status() === 200 && bytes) {
-          const data = bytes.get_data();
-          if (data && data.length > 0) {
-            const outFile = Gio.File.new_for_path(outputPath);
-            const parentDir = outFile.get_parent();
-            if (parentDir && !parentDir.query_exists(null)) {
-              GLib.mkdir_with_parents(parentDir.get_path(), 0o755);
-            }
-            outFile.replace_contents(data, null, false,
-              Gio.FileCreateFlags.REPLACE_DESTINATION, null);
-            if (callback) callback(true);
-            return;
-          }
-        }
-        if (callback) callback(false);
-      } catch (e) {
-        log(`Switcher – favicon fetch failed: ${e}`);
-        if (callback) callback(false);
-      }
-    });
-  } catch (e) {
-    log(`Switcher – favicon session failed: ${e}`);
-    if (callback) callback(false);
-  }
-}
-
-function fetchFaviconForProvider(provider, callback) {
-  const domain = extractDomain(provider.url);
-  if (!domain) {
-    if (callback) callback(false);
-    return;
-  }
-  const faviconUrl = `https://www.google.com/s2/favicons?domain=${domain}&sz=32`;
-  const outputPath = getIconPath(provider.keyword);
-  fetchFavicon(faviconUrl, outputPath, callback);
 }
 
 function buildWebSearchGroup(settings) {
